# atodo-api-spec.yaml
...
openapi: 3.0.3
info:
  title: A-To-Do API
  description: >-
    Backend API for the A-To-Do to-do list application. Implementing this
    spec turns the app from a single-browser, localStorage-only prototype
    into a real multi-tenant service: every endpoint below is scoped to the
    authenticated account (via its bearer token), so two people logging into
    two different accounts genuinely see two different task lists, rather
    than the one shared list the pre-backend prototype used.

    All endpoints are prefixed with `/atodo/v1`. The bearer token is an
    opaque, server-issued credential (a real signed JWT is the intended
    implementation) that embeds at least `sub` (the account id) and the
    account's current `subscription` claim, so the client can read
    subscription status without an extra round trip -- see the `Token`
    schema. Whenever an endpoint changes something the token's claims
    reflect (starting/cancelling/scheduling deletion of a subscription), it
    returns a freshly issued token in place of the old one; the client is
    expected to discard the old one immediately.

    Every error response uses the same `Error` schema (`{ code, message }`).
    `code` is a stable, machine-readable identifier the client branches on
    (e.g. to show a specific translated message); `message` is a
    human-readable fallback, not meant to be shown as-is in a finished UI.
  version: "1.0.0"
...
servers:
  - url: "{baseUrl}/atodo/v1"
    variables:
...
security:
  - bearerAuth: []
...
paths:
...
  /tasks:
    get:
      tags: [Tasks]
      summary: Fetch every task belonging to the current account
      description: >-
        Returns the account's entire task list in one call -- realistic
        for a personal to-do list's scale, and simpler than paginating a
        collection that's meant to all be visible/editable in one screen
        anyway.
      responses:
        "200":
          description: The account's tasks.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Task" }
        "401":
          description: Missing/invalid bearer token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    put:
      tags: [Tasks]
      summary: Replace the current account's entire task list
      description: >-
        Bulk replace, not a per-task PATCH/POST -- the client already
        computes the full resulting list locally for every mutation (adding
        a task, completing an occurrence, splitting a recurring series,
        ...) before saving, so this mirrors that instead of requiring a
        separate granular endpoint per mutation kind. `id` is
        client-generated (see the `Task` schema) and must be unique within
        the array; the server doesn't invent or reconcile ids.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items: { $ref: "#/components/schemas/Task" }
      responses:
        "200":
          description: Saved -- echoes back exactly what was stored.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Task" }
        "401":
          description: Missing/invalid bearer token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /subscriptions/trial:
    post:
      tags: [Subscriptions]
      summary: Start a 14-day Pro trial for the current account
      description: >-
        Identical to a paid Pro subscription while it lasts. Currently
        allows repeat trials (useful for testing); a production
        implementation is expected to allow only one trial per account
        ever, tracked server-side.
      responses:
        "200":
          description: Trial started.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { $ref: "#/components/schemas/Token" }
                  user: { $ref: "#/components/schemas/User" }
        "401":
          description: Missing/invalid bearer token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /subscriptions/checkout-sessions:
    post:
      tags: [Subscriptions]
      summary: Start a Stripe Checkout session for a Pro subscription
      description: >-
        Creates a Stripe Checkout session server-side (Stripe API keys
        never touch the client) and returns the URL to redirect the
        browser to. Stripe redirects back to `successUrl`/`cancelUrl`
        (this app's `success.html`/`cancel.html`) once the visitor finishes
        or abandons payment on Stripe's own hosted page; the client polls
        `GET /subscriptions/checkout-sessions/{sessionId}` (from
        `checkout.html`, the intermediate page shown while waiting) to find
        out which.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [billingInterval, successUrl, cancelUrl]
              properties:
                billingInterval:
                  type: string
                  enum: [monthly, annual]
                successUrl:
                  type: string
                  format: uri
                  description: Where Stripe redirects on completed payment (this app's success.html).
                cancelUrl:
                  type: string
                  format: uri
                  description: Where Stripe redirects if the visitor backs out (this app's cancel.html).
      responses:
        "200":
          description: Session created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId: { type: string }
                  checkoutUrl:
                    type: string
                    format: uri
                    description: Stripe's own hosted checkout page -- redirect the browser here.
        "401":
          description: Missing/invalid bearer token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /subscriptions/checkout-sessions/{sessionId}:
    get:
      tags: [Subscriptions]
      summary: Poll a checkout session's status
      description: >-
        `checkout.html` polls this until `status` is no longer `pending`.
        Once `paid`, `token`/`user` reflect the now-active Pro
        subscription, same shape as every other subscription-mutating
        endpoint's response, ready to replace whatever's currently stored
        client-side.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Current status.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [pending, paid, cancelled]
                  token: { $ref: "#/components/schemas/Token" }
                  user: { $ref: "#/components/schemas/User" }
        "401":
          description: Missing/invalid bearer token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: No such checkout session (for this account).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /subscriptions/cancel:
    post:
      tags: [Subscriptions]
      summary: Cancel the current subscription at period end
      description: >-
        Stops future renewal; access/limits are untouched until
        `expiresAt`, same as a real "cancel at period end" would behave.
      responses:
        "200":
          description: Cancelled.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { $ref: "#/components/schemas/Token" }
                  user: { $ref: "#/components/schemas/User" }
        "401":
          description: Missing/invalid bearer token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: No active subscription to cancel.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { code: NO_ACTIVE_SUBSCRIPTION, message: "There's no active subscription to cancel." }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        `Authorization: Bearer <token>`, as returned by `POST /auth/login`
        or any of the subscription-mutating endpoints. See the `Token`
        schema for what it's expected to carry.
  schemas:
    EpochMillis:
      type: integer
      format: int64
      description: >-
        Milliseconds since the Unix epoch (`Date.now()`/`new Date(ms)` on the
        client), not an RFC 3339 string -- deliberately, since every one of
        these fields is used in client-side arithmetic/sorting (elapsed
        time, "is this still active", "which of these was created first",
        ...) and epoch millis need no parsing to do that. `dueDate`/
        `endDate`/occurrence-date fields elsewhere are a different thing --
        plain `YYYY-MM-DD` strings with no time component, compared as
        strings, so those stay `format: date`.
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable, machine-readable -- the client branches on this, not `message`.
        message:
          type: string
          description: Human-readable fallback; a finished client should show its own translated copy per `code` instead.
    Token:
      type: string
      description: >-
        Opaque to the client beyond reading its claims (a real
        implementation: a signed JWT). Expected claims: `sub` (the account
        id) and `subscription` (the same shape as `User.subscription`, a
        snapshot as of when the token was issued -- re-issued by every
        endpoint that changes it, see each one's own description). The
        client never trusts a locally decoded token's claims for anything
        but quick, non-authoritative UI reads (e.g. "should the Subscribe
        button show right now") -- `GET /auth/me` is the actual source of
        truth for whether the account/subscription are still genuinely in
        that state.
    Background:
      type: object
      nullable: true
      description: The Settings background picker's currently-applied Unsplash photo, or null for none.
      properties:
        regularUrl: { type: string, format: uri }
        thumbUrl: { type: string, format: uri }
        photographerName: { type: string }
        photographerUrl: { type: string, format: uri }
        photoLink: { type: string, format: uri }
    Subscription:
      type: object
      nullable: true
      description: null means the account has never subscribed (the "Free" plan).
      properties:
        id: { type: string }
        plan:
          type: string
          enum: [trial, pro]
        billingInterval:
          type: string
          enum: [monthly, annual]
          nullable: true
          description: null for a trial.
        startedAt: { $ref: "#/components/schemas/EpochMillis" }
        expiresAt:
          allOf: [{ $ref: "#/components/schemas/EpochMillis" }]
          description: >-
            The single source of truth for whether the subscription is
            currently active -- always `now < expiresAt`, never a separately
            cached boolean, so an expiry is correctly detected the instant
            it passes rather than only once something happens to re-check it.
        cancelAtPeriodEnd:
          type: boolean
          description: True once the subscription won't renew (see POST /subscriptions/cancel and POST /users/me/schedule-deletion) -- access is untouched until expiresAt regardless.
        scheduledDeletion:
          type: boolean
          description: True if the account itself, not just the subscription, is set to be deleted once expiresAt passes -- see POST /users/me/schedule-deletion.
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        nickname: { type: string }
        avatar:
          type: string
          nullable: true
          description: A data URL, or null for the initials-only fallback avatar.
        timeFormat: { type: string, enum: ["12", "24"] }
        background: { $ref: "#/components/schemas/Background" }
        language:
          type: string
          enum: [en, hr]
          nullable: true
          description: null means never chosen or IP-detected yet -- the client detects and PATCHes a real value the first time this is null right after login.
        activeTaskId:
          type: string
          nullable: true
          description: Which task (by Task.id) is currently focused/being timed, if any.
        activeOccurrenceDate:
          type: string
          format: date
          nullable: true
          description: Which occurrence of activeTaskId, for a recurring task that can show more than one row at once. Always null when activeTaskId is null.
        todoViewMode:
          type: string
          enum: [pending, next-recurrence, all]
        subscription: { $ref: "#/components/schemas/Subscription" }
    FrequencyOnce:
      type: object
      required: [type]
      properties:
        type: { type: string, enum: [once] }
    FrequencyDays:
      type: object
      required: [type, interval]
      properties:
        type: { type: string, enum: [days] }
        interval: { type: integer, minimum: 1 }
    FrequencyWeeks:
      type: object
      required: [type, interval]
      properties:
        type: { type: string, enum: [weeks] }
        interval: { type: integer, minimum: 1 }
        weekdays:
          type: array
          items: { type: integer, minimum: 0, maximum: 6 }
          description: 0=Sunday..6=Saturday. Omitted/empty falls back to dueDate's own weekday (backward compatibility with older saved tasks).
    FrequencyMonths:
      type: object
      required: [type, interval]
      properties:
        type: { type: string, enum: [months] }
        interval: { type: integer, minimum: 1 }
        dayMode:
          type: string
          enum: [day, last, before-last, weekday, multi-weekday, multi-weekday-offset, multi-day]
          description: Omitted/"day" anchors to dueDate's day-of-month, clamped to shorter months.
        offset:
          type: integer
          minimum: 0
          maximum: 3
          description: dayMode "before-last" only -- days before the last day of the month.
        weekday:
          type: integer
          minimum: 0
          maximum: 6
          description: dayMode "weekday" only.
        ordinal:
          oneOf:
            - type: integer
              minimum: 1
              maximum: 5
            - type: string
              enum: [last]
          description: dayMode "weekday"/"multi-weekday"/"multi-weekday-offset". "last" only valid for dayMode "weekday".
        weekdays:
          type: array
          items: { type: integer, minimum: 0, maximum: 6 }
          description: dayMode "multi-weekday"/"multi-weekday-offset" only.
        offsetDirection:
          type: string
          enum: [before, after]
          description: dayMode "multi-weekday-offset" only.
        offsetDays:
          type: integer
          minimum: 0
          maximum: 6
          description: dayMode "multi-weekday-offset" only.
        days:
          type: array
          items: { type: integer, minimum: 1, maximum: 31 }
          description: dayMode "multi-day" only -- exact days of the month.
    Frequency:
      description: See recurrence.js's own top-of-file comment for the full recurrence rules each shape implies.
      oneOf:
        - $ref: "#/components/schemas/FrequencyOnce"
        - $ref: "#/components/schemas/FrequencyDays"
        - $ref: "#/components/schemas/FrequencyWeeks"
        - $ref: "#/components/schemas/FrequencyMonths"
      discriminator:
        propertyName: type
        mapping:
          once: "#/components/schemas/FrequencyOnce"
          days: "#/components/schemas/FrequencyDays"
          weeks: "#/components/schemas/FrequencyWeeks"
          months: "#/components/schemas/FrequencyMonths"
    Timer:
      type: object
      nullable: true
      properties:
        mode: { type: string, enum: [countdown, countup] }
        continuePastZero: { type: boolean }
        totalSeconds: { type: integer }
        remainingSeconds:
          type: integer
          description: The checkpointed remaining time as of runningSince (or the frozen value if paused) -- the client derives live remaining time from this plus elapsed wall-clock time, not a stored live countdown.
        runningSince:
          allOf: [{ $ref: "#/components/schemas/EpochMillis" }]
          nullable: true
          description: null while paused/not yet started.
        occurrenceDate:
          type: string
          format: date
          description: Which occurrence of the task this timer belongs to.
    LogEntry:
      type: object
      properties:
        message: { type: string }
        timestamp: { $ref: "#/components/schemas/EpochMillis" }
        occurrenceDate:
          type: string
          format: date
          nullable: true
    Comment:
      type: object
      properties:
        text: { type: string }
        timestamp: { $ref: "#/components/schemas/EpochMillis" }
    Task:
      type: object
      required:
        - id
        - taskId
        - seriesId
        - name
        - dueDate
        - allDay
        - appointment
        - passive
        - recurUntilCompleted
        - frequency
        - completions
        - dismissed
        - markedFailed
        - createdAt
      properties:
        id:
          type: string
          description: This specific record's id -- changes across a recurring-series split (see taskId).
        taskId:
          type: string
          description: >-
            Stable across a recurring task's whole lineage, including every
            fragment a "this occurrence"/"this and following" edit or delete
            has ever split off from it (see id) -- this, not id, is "one
            task" for any UI or limit that means the logical task a person
            thinks of, not one specific record.
        seriesId:
          type: string
          description: Groups otherwise-unrelated tasks a person has manually merged into one series in the Manage Tasks modal -- independent of taskId.
        seriesName:
          type: string
          nullable: true
        name: { type: string }
        description: { type: string }
        details: { type: string }
        dueDate: { type: string, format: date }
        dueTime:
          type: string
          nullable: true
          description: "HH:MM, 24-hour. null when allDay."
        allDay: { type: boolean }
        appointment:
          type: boolean
          description: Past-due becomes permanently "failed" instead of carrying over as overdue.
        passive:
          type: boolean
          description: A plain reminder with no "done" state -- its checkbox means "mark failed" (see markedFailed) instead of complete.
        recurUntilCompleted: { type: boolean }
        pendingReschedules:
          type: array
          items: { type: string, format: date }
          description: recurUntilCompleted tasks only -- every date its current occurrence has been auto-pushed to since dueDate, still-live until the chain is finally resolved.
        endDate:
          type: string
          format: date
          nullable: true
        frequency: { $ref: "#/components/schemas/Frequency" }
        completions:
          type: object
          additionalProperties: { type: boolean }
          description: Map of occurrence date (ISO) -> true.
        dismissed:
          type: object
          additionalProperties: { type: boolean }
        markedFailed:
          type: object
          additionalProperties: { type: boolean }
        createdAt:
          allOf: [{ $ref: "#/components/schemas/EpochMillis" }]
          description: Of this taskId's lineage, not this specific record -- carried forward unchanged across every split/manual-occurrence fragment (see the free-tier task limits, which grandfather by this field).
        timer: { $ref: "#/components/schemas/Timer" }
        log:
          type: array
          items: { $ref: "#/components/schemas/LogEntry" }
        comments:
          type: array
          items: { $ref: "#/components/schemas/Comment" }